+++ src/main/java/com/mraof/minestuck/alchemy/GristGutter.java
@@ -6,6 +6,7 @@ import com.mraof.minestuck.player.IdentifierHandler;
 import com.mraof.minestuck.player.PlayerData;
 import com.mraof.minestuck.player.PlayerIdentifier;
 import com.mraof.minestuck.player.PlayerSavedData;
+import com.mraof.minestuck.skaianet.SburbPlayerData;
 import com.mraof.minestuck.skaianet.Session;
 import com.mraof.minestuck.skaianet.SessionHandler;
 import net.minecraft.nbt.EndTag;
@@ -17,11 +18,12 @@ import net.minecraft.util.RandomSource;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.server.ServerLifecycleHooks;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * A class that handles Grist overflow whenever you acquire too much grist.
@@ -34,19 +36,22 @@ public class GristGutter
 	
 	public static final int GUTTER_CAPACITY = 10000;
 	
+	private final MinecraftServer mcServer;
 	private final Session session;
 	private final NonNegativeGristSet gristSet;
 	private long gristTotal;
 	
-	public GristGutter(Session session)
+	public GristGutter(MinecraftServer mcServer, Session session)
 	{
+		this.mcServer = mcServer;
 		this.session = session;
 		this.gristSet = new NonNegativeGristSet();
 		this.gristTotal = 0;
 	}
 	
-	public GristGutter(Session session, Tag tag)
+	public GristGutter(MinecraftServer mcServer, Session session, Tag tag)
 	{
+		this.mcServer = mcServer;
 		this.session = session;
 		this.gristSet = NonNegativeGristSet.CODEC.parse(NbtOps.INSTANCE, tag).resultOrPartial(LOGGER::error).orElse(new NonNegativeGristSet());
 		this.gristTotal = 0;
@@ -79,6 +84,11 @@ public class GristGutter
 		return gristSet.asImmutable();
 	}
 	
+	private Stream<PlayerIdentifier> gutterPlayers()
+	{
+		return this.session.getPlayers().stream().filter(player -> SburbPlayerData.get(player, mcServer).hasEntered());
+	}
+	
 	public long getRemainingCapacity()
 	{
 		return Math.max(0, (long) (GUTTER_CAPACITY * gutterMultiplierForSession()) - gristTotal);
@@ -86,14 +96,11 @@ public class GristGutter
 	
 	public double gutterMultiplierForSession()
 	{
-		PlayerSavedData playerSavedData = PlayerSavedData.get(ServerLifecycleHooks.getCurrentServer());
-		double gutterMultiplier = 0;
-		for(PlayerIdentifier player : this.session.getPlayers())
-		{
-			PlayerData data = playerSavedData.getData(player);
-			gutterMultiplier += data.getGutterMultipler();
-		}
-		return gutterMultiplier;
+		PlayerSavedData playerSavedData = PlayerSavedData.get(mcServer);
+		
+		return this.gutterPlayers()
+				.map(player -> playerSavedData.getData(player).getGutterMultipler())
+				.reduce(0D, Double::sum);
 	}
 	
 	/**
@@ -164,27 +171,25 @@ public class GristGutter
 		if(event.phase == TickEvent.Phase.START && event.getServer().overworld().getGameTime() % 200 == 0)
 		{
 			for(Session session : SessionHandler.get(event.getServer()).getSessions())
-			{
-				session.getGristGutter().distributeToPlayers(session.getPlayers(), event.getServer());
-			}
+				session.getGristGutter().distributeToPlayers();
 		}
 	}
 	
-	private void distributeToPlayers(Set<PlayerIdentifier> players, MinecraftServer server)
+	private void distributeToPlayers()
 	{
-		RandomSource rand = server.overworld().random;
-		List<PlayerIdentifier> playerList = new ArrayList<>(players);
+		RandomSource rand = mcServer.overworld().random;
+		List<PlayerIdentifier> playerList = this.gutterPlayers().collect(Collectors.toCollection(ArrayList::new));
 		Collections.shuffle(playerList, new Random(rand.nextLong()));
 		
 		for(PlayerIdentifier player : playerList)
 		{
-			tickDistributionToPlayer(player, server, rand);
+			tickDistributionToPlayer(player, rand);
 		}
 	}
 	
-	private void tickDistributionToPlayer(PlayerIdentifier player, MinecraftServer server, RandomSource rand)
+	private void tickDistributionToPlayer(PlayerIdentifier player, RandomSource rand)
 	{
-		PlayerData data = PlayerSavedData.getData(player, server);
+		PlayerData data = PlayerSavedData.getData(player, mcServer);
 		
 		long spliceAmount = (long) (data.getEcheladder().getGristCapacity() * getDistributionRateModifier());